+++ r2-testapp/src/main/java/org/readium/r2/testapp/epub/EpubActivity.kt
@@ -47,7 +47,7 @@ import org.readium.r2.testapp.reader.BookData
 import org.readium.r2.testapp.reader.EpubReaderFragment
 import org.readium.r2.testapp.reader.ReaderViewModel
 import org.readium.r2.testapp.reader.toNavigatorHighlight
-import org.readium.r2.testapp.utils.NavigatorContract
+import org.readium.r2.testapp.reader.ReaderActivityContract
 import java.lang.IllegalStateException
 
 class EpubActivity : R2EpubActivity() {
@@ -72,7 +72,7 @@ class EpubActivity : R2EpubActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         check(LibraryActivity.isServerStarted)
 
-        val inputData = NavigatorContract.parseIntent(this)
+        val inputData = ReaderActivityContract.parseIntent(this)
         val publication = inputData.publication
         val bookId = inputData.bookId
         val baseUrl = requireNotNull(inputData.baseUrl)
