+++ r2-testapp/src/main/java/org/readium/r2/testapp/utils/FlowLifecycleObserver.kt
@@ -10,8 +10,13 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.flow.collect
 
-// See https://proandroiddev.com/android-singleliveevent-redux-with-kotlin-flow-b755c70bb055
+/*
+ * Copyright 2021 Readium Foundation. All rights reserved.
+ * Use of this source code is governed by the BSD-style license
+ * available in the top-level LICENSE file of the project.
+ */
 
+// See https://proandroiddev.com/android-singleliveevent-redux-with-kotlin-flow-b755c70bb055
 
 inline fun <reified T> Flow<T>.observeWhenStarted(
     lifecycleOwner: LifecycleOwner,
